consistent use of super()
