+++ plugin/flags_sources/flags_file.py
@@ -34,7 +34,7 @@ class FlagsFile(FlagsSource):
             include_prefixes (str[]): A List of valid include prefixes.
             search_scope (SearchScope): Where to search for a flags file.
         """
-        super(FlagsFile, self).__init__(include_prefixes)
+        super().__init__(include_prefixes)
         # TODO(igor): Do we need a search scope here?
         self.__search_scope = search_scope
 
@@ -47,9 +47,9 @@ class FlagsFile(FlagsSource):
         Returns: str[]: Return a list of flags in this .clang_complete file
         """
         log.debug(" [clang_complete_file]: for file %s", file_path)
-        cached_flags_path = FlagsFile.get_cached_from(file_path)
+        cached_flags_path = super().get_cached_from(file_path)
         log.debug(" [clang_complete_file]:[cached]: '%s'", cached_flags_path)
-        flags_file_path = FlagsFile.find_current_in(
+        flags_file_path = super().find_current_in(
             self.__search_scope)
         log.debug(" [clang_complete_file]:[current]: '%s'", flags_file_path)
 
@@ -71,35 +71,6 @@ class FlagsFile(FlagsSource):
                 FlagsFile.path_for_file[file_path] = flags_file_path
         # now we return whatever we have
         return flags
-
-    @classmethod
-    def get_cached_from(cls, file_path):
-        """Get cached path for file path.
-
-        Args:
-            file_path (str): Input file path.
-
-        Returns:
-            str: Path to the cached flag source path.
-        """
-        if file_path and file_path in cls.path_for_file:
-            return cls.path_for_file[file_path]
-        return None
-
-    @classmethod
-    def find_current_in(cls, search_scope):
-        """Find current path in a search scope.
-
-        Args:
-            search_scope (SearchScope): Find in a search scope.
-
-        Returns:
-            str: Path to the current flag source path.
-        """
-        return File.search(
-            file_name=cls._FILE_NAME,
-            from_folder=search_scope.from_folder,
-            to_folder=search_scope.to_folder).full_path()
 
     def __flags_from_clang_file(self, file):
         """Get flags from .clang_complete file.
